+++ packages/AppearanceOfEditor/Base.js
@@ -1,0 +1,4 @@
+const color = require('../BaseColor')
+module.exports = {
+  "focusBorder": color['blue_ctrl'],
+}