+++ packages/AppearanceOfEditor/Widget.js
@@ -1,0 +1,5 @@
+const color = require('../BaseColor')
+module.exports = {
+     // find/replace
+     "editorWidget.background": color['black'],
+}